Prepare to use HAML and Draper
Added gems for HAML (as template language instead of ERB) and Draper (decorator for models).
Started converting ERB views to HAML.
Decided to start slowly removing http parameter filtering from models (attr_accesible), because it is wrong: they should be filtered in the controllers (using before_filter, for example), or in some kind of a "conductor".
Note: decorators/presenters and conductors are additional layers to MVC design.